Increase device update interval to 15 minutes
The device update interval is a 'fallback', and will not be used when
the device is active. In that case, the websocket updates will update
the device much faster.

When the device is not active, most fetches will result in retrieving
the same data for an offline device. Therefore, once every 15 minutes
should be more than enough.